Fix check for not finding open collections
Long back when Magma refactored its status code to include a new one
called as OkDocNotFound, which indicates that fetching a doc was
successful but the doc doesn't exist, we missed refactoring the check at
one place. This patch fixes that check.

Here's the old patch where the refactoring was originally done:
https://review.couchbase.org/c/kv_engine/+/201289

Change-Id: I1025e669bce72be2200260db4c923f62cd21e51c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229247
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>