MB-67254: Improve the robustness for AccessScanner test
The access scanner is scheduled as a task to run at a given
time. It does *not* run the visitors itself, instead the task
schedules a number of tasks (one per shard) to visit the
hashtable and create the access log files.

This means that in order to wait for the access scanner to complete a
full run we need to wait until all of these tasks complete. (which
would be the sum of ep_num_access_scanner_skips and
ep_num_access_scanner_runs and that the AccessScanner task itself
is snoozed).

Change-Id: I86e4b3aa19d95bca5047d92d5ad0be27334aa272
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229483
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>