MB-67207: Remove the ability to set-with-meta to replica vbuckets
The FORCE_WITH_META_OP set/add/del_with_meta flag permits a mutation
against pending and replica vbuckets. This is unsafe and will cause
such problems and crashes as seen in MB-67207 where a replica will
get out of sync with DCP.

This commit keeps the option, but removes the replica/pending write
aspect, making it equivalent to SKIP_CONFLICT_RESOLUTION_FLAG.

The flag is retained in the protocol as there could be some legitimate
case out there where this is successfully used against active vbuckets.
In that case, the flag is safe (and just skips conflicts).

Change-Id: I6dfca7351c2c8a6ebf3bc016e96c6127952ad573
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229489
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>