MB-67263: Hold VB state lock in read-mode for entire streamRequest
The VB state lock is acquired in a few places in streamRequest, but
not broad enough to prevent unusual inconsistencies with VB state
changes - e.g. a streamRequest could race with a failover, observe
state is active, but not a new failover entry that is generated
under the state lock.

E.g. in the observed field issue, an indexing stream connected and
requested active only...

streamRequest could pass the state check but overlap with the
failover setVbucketState. The stream request doesn't observe the
failover entry added for failover!

T0 setVbState:

  // obtain exclusive lock on VB state....
  folly::SharedMutex::WriteHolder vbStateLock(vb->getStateLock());
  ...
  // switch the vbucket state (now active)
  auto oldstate = vbMap.setState_UNLOCKED(*vb, to, meta, vbStateLock);

T1: streamRequest
  // Sees state is active and continues. Use of the state lock would
  // stop the request here
  if ((flags & DCP_ADD_STREAM_ACTIVE_VB_ONLY) &&
      (vb->getState() != vbucket_state_active)) {
      logger->info(
              "({}) Stream request failed ...");
      return cb::engine_errc::not_my_vbucket;
  }
  ...
  // Do rollback stuff and get failover table
  auto rollbackResult = doRollbackCheck(*vb, ...);
  auto failoverEntries =
            std::get<std::vector<vbucket_failover_t>>(rollbackResult);

T0: setVbState
  // Now the failover entry is created.
  if (to == vbucket_state_active ...) {
      ....
      vb->failovers->createEntry(highSeqno);

Change-Id: Iacae11eeff765f9352cc41a7024ba22769756d02
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229545
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker