MB-66810: Fix failing unit tests
Previous code for MB-66810 was tested before the addition of
the ProcessStreamsExceptionDisconnectsConnection test.

The MB-66810 changes now break that test as that test was
relying on set-vb-state items, which are no longer queued
for ephemeral.

Write one extra item so that the exception handling can be
tested.

Change-Id: I1c045664958fc398f0de4a5167d4f456c09daff1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229603
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>