MB-67283: Add utilities/fusion_support.h to reduce code complexity
The memcached core gets compiled with (EE) and without (CE)
magma available, and fusion is only supported on a subset
of EE platforms.

To avoid cluttering the code with ifdef's making the code
harder to read we hide the difference in a single file
and provide a stub for the magma functions so that we
don't need a preprocessor macro.

Change-Id: Ia101355befcd8c5428ed1a489598380af413b967
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229523
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>