MB-66958: Add timeout for abrupt shutdown
Terminate the process with exit code 9 after a configurable amount of
time when stdin close. The timeout is specified in the configuration
file memcached.json under the key abrupt_shutdown_timeout which
is specified as a number in milliseconds.

If the variable is absent (or a value of 0) disable the timeout.
This is the default behavior.

Change-Id: Ibf99fbfb973617ab12c68bd24f897573266e0b89
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/229746
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>