MB-67411: Fix error code for IPV6_V6ONLY on WIN32
One should call WSAGetLastError and not use errno when
setsockopt fails on windows (it is part of winsock).

While at it we might as well remove the #ifdef as it
is supported on all our supported platforms to make
it clear that the code isn't optional.

Tweak the error messages to print the option as a separate
element in the context.

Change-Id: I92cdeddfe70b4f4f309a10638b98a5d52d587c23
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230011
Tested-by: Trond Norbye <trond.norbye@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>