MB-67447: Validate scope exists when adding a collection
Prior to this a DCP stream wasn't validating a scope exists
allowing for KV to create a replica with an inconsistent collection
state. In earlier releases when we had a scope.data_size counter,
we would crash in the flusher if a mutation occurs against collections
in the scope (or even when accounting for collection system events of
the scope, this was in 7.1.x to 7.6.x)  with this change we now fail
early, preventing a replica being built and used in an inconsistent
state.

Change-Id: I662a4531b93f17edb1f21282e91cb95b812dd446
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230110
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>