[BP] MB-67447: Validate scope exists when adding a collection
Prior to this a DCP stream wasn't validating a scope exists
allowing for KV to create a replica with an inconsistent collection
state. In earlier releases when we had a scope.data_size counter,
we would crash in the flusher if a mutation occurs against collections
in the scope (or even when accounting for collection system events of
the scope, this was in 7.1.x to 7.6.x)  with this change we now fail
early, preventing a replica being built and used in an inconsistent
state.

(cherry picked from commit d07943b5256f82204946cb9286ce1592c7b01be6)

Change-Id: Idc445a03b7a057ec01dbcda0aadc2adff60ff614
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230882
Well-Formed: Restriction Checker
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>