MB-67546: Don't use copy_if_different for code generation steps
For `ADD_COMMAND()`, it is important that the specified commands
actually generate the `OUTPUT` files; otherwise CMake will still believe
the output files are out of date and re-generate them every time you
build, and then re-run any build steps which depend on those files. As
such, using `copy_if_different` to copy the generated files to the
expected OUTPUT location is not desirable.

Change-Id: I3c8f611ffc73f8a92176531f2702222932ef7a4b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230421
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>