MB-66959 dbfill: Add option to send mutations to selected vbucket
Its useful to have a tool that can load data into a specific vbucket
only. dbfill is well suited to offer that as it directly understand mcbp
and can talk to it.

Add an option to dbfill that accepts the target vbucket into which mutations will be performed.

./cbstats localhost:11210 -u Administrator -p password -b default vbucket-details 1 | grep high_seqno
vb_1:high_seqno:                          20290

./dbfill --bucket default -h localhost -u Administrator -P password --num-items 100 --vbucket 1
       100 stored          0 tmpfail    0 failed    100 ops/s - done
 - done

./cbstats localhost:11210 -b default -u Administrator -p password vbucket-details 1 | grep high_seqno
vb_1:high_seqno:                          20390

Change-Id: I7107a7410ac4a359dba4457c3df86cc9a033768e
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230487
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>