MB-67574: Fix exclusive end key computation
A range scan with the range specified as exclusive at the end will set
the end key for the scan, to the specified end key with the last byte
decremented by 1 and with std::numeric_limits<char>::max() appended.
This is \x7F on x86 Linux and \xFF on ARM Linux. This is because the
char type is not defined to be signed/unsigned by the standard, and the
implementation differs.

The \x7F is within the range of valid UTF-8 byte sequences.

Assume we have end="user:;" and exclusive = true. This results in an
inclusive end of "user::\x7d".

  user::Alice = user::\x41\x6C\x69\x63\x65
  user::Алекс = user::\xd0\x90\xd0\xbb\xd0\xb5\xd0\xba\xd1\x81

The second key starts at \xd0 > \x7d and falls outside the range, so it
will be skipped.

Change-Id: I458cfe11cf48cd68a45e42ce15cfe4e875053dca
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230508
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker