[jenkins_failures.py] Extract NODE_NAME from console output
The node name is the container's name which can be linked to the VM and
then to the physical machine.

This will extract and print the node's name to link commonly failing
tests to the machine they run on. This can be used to deduce if a
particular machine may be experience slow hardware issues.

Change-Id: I1d2132fa953e4e8d214ea438a80042e5c647cfdc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230557
Tested-by: Faizan Alam <faizan.alam@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>