MB-66831: Send v2.0 snap marker if HPS out side snapshot range
There was a minor bug introduced via MB-67418, where we are enqueueing
a SnapshotMarker, with purge seqno: 0 & HPS: std::nullopt; in the
snapshot codec logic, if either of them has a value, we send out
snapshot marker v2.2

In this particular test:

- There is a offline upgrade of the cluster from 7.1 to 8.0.
- The replica has HPS on disk set to a non-zero value (due to older
  logic where we bumped the HPS to snapEnd each time entire snapshot is
  consumed, despite not having any durability work load).
- When the 8.0 cluster starts, since both the consumer & producer have
  negotiated marker_version=2.2, active sends out HPS=0 & replica has
  warmed up from disk a non-zero value it cause a monotonic violation on
  the replica.

Note: This is currently a temporary fix to get past the crashes &
unblock testing. Discussion still in progress if we want to leave this
fix as is or come-up with other alternate ways of fixing this.

Change-Id: I867e8f0eeff23071262de249bb5befcf70de46e2
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230597
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>