MB-66831: Ignore producer sent HPS outside snapshot range
If this cluster was offline upgraded from < 8.0 to 8.0, the HPS in the
PDM state loaded during warmup could be ahead of the HPS sent from the
active.

Consider the following scenario:
- There is a offline upgrade of the cluster from 7.1 to 8.0.
- The replica has HPS on disk set to a non-zero value (due to older
  logic where we bumped the HPS to snapEnd each time entire snapshot is
  consumed, despite not having any durability work load).
- When the 8.0 cluster starts, since both the consumer & producer have
  negotiated marker_version=2.2, active sends out HPS=0 & replica warms
  up from disk a non-zero value, which cause a monotonic violation on
  the replica, eventually when the active sent snapshot marker is
  processed.

If the HPS is not in the snapshot range, pretend the HPS was never sent
in the snapshot marker on the consumer.

Also revert some part of the test changes made in ep_testsuite_dcp as
part of https://review.couchbase.org/c/kv_engine/+/225176, which after
the new changes to update the HPS on the PDM didn't send back a
SeqnoAck.

Change-Id: I2fd68d3094a621e0ce08f0794c1b399b5bf7d121
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230769
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>