MB-67520: Define item expiry/locking as 32-bit
Documents are given an absolute expiry time and the persisted metadata
only has a 32-bit field for storing that value. Similarly the
hash-table only reserves 32-bits for storing the same. The same is for
deleted documents, where the "delete-time" value is stored in the same
4-bytes. Similarly the lock timeout is also restricted to 4-bytes.

This commit removes the use of time_t when we are dealing with document
expiry/delete-time and lock timeouts and uses the explicitly sized
uint32_t type so it's clear that 32-bits are permitted for expiry
(instead of more sugary named types).

In terms of narrowing, the code uses gsl::narrow_cast and accepts that
there is minuscule risk that time could be >2^32. MB-67520 aims to
primarily clean-up and get a reasonable chunk of code free of narrowing
warnings and will defer better handling of the remote possibility of
>2^32 time to later releases.

Change-Id: I8b7e612dbd890fc621de0df6ef222ab25bfc5677
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231203
Reviewed-by: Mohammad Zaeem <mohammad.zaeem@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>