MB-67520: Fix many narrowing warnings in dcp
DCP code is mostly mixing u32/u64 around flow control, the only part
of flow control that is limited to u32 is the protocol message for
acking bytes (defines a u32 field). Internally we should use size_t
throughout, e.g. getMessageSize returns the size of the message and
that doesn't have to be limited to u32. Flow control counters are
already 64 (now size_t) and we could of in theory accumulated more
data in the consumer that we can ack in one message - this "overflow"
is now checked against in flow-control.cc, but the ack trigger
boundary is far far lower than 4billion.

Change-Id: I7521af79569009b28e2757e2a1a73fe3ed9aaa1c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230725
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>