MB-67545: Fix intermittent GetRandomKey_access failure
These tests can fail when the flusher has not yet completed. The
GetRandomKey operation attempts to find a random key from a collection.
However, it may fail even if such a key exists. This happens because
GetRandomKey relies on collection stats, which are only updated after
the flusher has finished. So, if a flush is still in progress, the stats
will not yet reflect the presence of items in the collection, causing
GetRandomKey to incorrectly determine that no valid key is available.

Change-Id: I8f9e40ad2b28fa6be09d4666980a3c0cf70ddf42
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230751
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>