MB-67520: Change mc_time expiry limit method to use uint32_t
The expiry limiting code exists to take an already calculated item
expiry and check it doesn't exceed some defined limit, capping it
to the limit or changing from 0 to the limit as required.

The types are confused in that we are using time_t (64-bit) extensively
for general absolute epoch time. This is correct as that value can and
should be represented by the maxium unsigned integer type. However as
the document expiry time (also absolute epoch time) is limited to
4-bytes by the document metadata definition, time_t is incorrect in
this area.

This commit tidies this path up so that a uint32_t is the input and
output of the mc_time limiting finction (and server API wrapper). This
is a stepping stone to further cleaning up the u32/u64
expiry-time/system-time mismatch and silent narrowing.

Change-Id: I44ec1575bae8b48a1036f27ab806ea3280f32ff5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230967
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>