MB-67520: Add a function for generating document expiry time
Remove duplicated code that converts the protocol provided expiry
into a document expiry time and ensure a 32-bit type is used on the
on the result.

Later this function can be adapted to provide overflow detection in
the case when MCBP relative expiry results in a value that exceeds
u32 storage (which could occur if the server was started and reads
a system time > 2^32 seconds).

Change-Id: I5bbc539500e96c19503cff6e4347ac0be74939e1
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230968
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>