MB-67632: Fix persistence of snapEnd as HPS when v2.0 marker is used
The pre-Morpheus behaviour on replicas is to set the HPS to the seqno of
prepare for memory snapshots and to the snapEnd seqno for disk snapshots
(since disk snapshots send completed prepares as mutations), and we
cannot know the correct HPS there.

With the new changes and addition of v2.2 marker, we send the HPS from
the active.

The code preserved the branch matching the old behaviour of setting the
HPS to snapEnd, but due to other refactoring, that branch was never
executed. This is because the branch checks if the snapshot contains
nullopt for the HPS, but due to the restructuring, we always had a value
in the HPS of the snapshot range, only it was not correct for v2.0 disk
snapshots.

This issue did not affect the PDM seqno, since that logic is separate.

Since checkpoints now always track the HPS, I've moved the logic of
setting snapEnd to the checkpoint code, when the snapEnd seqno is
processed, we have all the information we need to determine what the HPS
for the snapshot is. This simplifies the flusher code.

The added tests check the persisted seqno is accurate to the PDM value,
but I have disabled the tests on Ephemeral until the follow up patches
which will fix it and re-enable.

Finally, the changes in dcp_stream_test.cc are partial reverts of change
https://review.couchbase.org/c/kv_engine/+/216147, since the tests were
changed to match the new faulty behaviour.

Change-Id: Iee6572dff7c5ce37b448fb1c31ffd2674451d3dc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230804
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>