MB-67633: Update SeqList HPS at the end of snapshot
When we finish processing a snapshot, we should update the HPS with the
value received from the active, or the snapEnd if no prepare was
received. This logic is equivalent to what the PDM does and what
persistent buckets do, and it ensures that if this replica is promoted,
it can correctly send the HPS to a replica (since we use the value from
the seqlist not the PDM in replication).

Change-Id: I5a141b3a9b94b64c7e700c4c7f865938808f9838
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/230805
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>