[jenkins_failures.py] Extract container's labels
The labels include info such as the VM's name that can be used to infer
details about the failing run.

Linux CVs run within containers which are run on VMs which are then run
on physical hardware. Node names rotate and so we can lose which VM
actually ran a particular run. Knowing the VM name therefore is more
useful to finding if particular hardware is degrading.

Change-Id: Ic6374d918e776c1bc49b8ab7641eef61304ef58d
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231135
Tested-by: Faizan Alam <faizan.alam@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>