MB-66315: Add PassiveStreamFuzzTest
Added a PassiveStream fuzz test which exercises receiving an initial
snapshot, which could be a disk or memory snapshot from an active.

Generates a SnapshotMarker with optional HPS/HCS/MVS/PurgeSeqno and
suitable mutations.

The test ensures that the input (marker + mutations) are logically
valid (in range, etc), and then funnels these into the PassiveStream.

The test also optionally flushes the checkpoint to disk.

Finally, it validates that the expected number of items was flushed and
the seqnos were updated to the expected value.

Change-Id: I4a44fbdb84f6d668b22dd9711fd18c52b2bc2212
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231283
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>