MB-67796: Always triggerCallback when the connection is shutting down
The original fix for MB-66385 did not account for the case where:
 - the connection is reset
 - the connection has a non-empty sendQueue
 - the connection has a reserved cookie (DCP connection / EWB)

In the DCP case, the ep-engine will destroy the DCP producer after the
conection is reset and call cookie.release() to drop the reference,
which is the last thing holding the Connection in the pending_close.

However, in MB-66385 we made disabled some notifications when there is
still data to send, and this affects the cookie.release() path too,
since in the current logic, if we have bytes to send, we do not trigger
the connection to fully close.

We need to either use force=true in the release() method, or (what this
patch does) we can simply not allow notifications to be missed during
connection closing.

Change-Id: I00564933f4d9914756bbdc90e2c916df2d6a076a
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231385
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Well-Formed: Restriction Checker