MB-67816: Update purgeSeqno even when equal to snapStartSeqno
We only updated the purgeSeqno if it is within the snapshot range, but
not the first seqno.

I am not 100% sure how likely this is to occur in the wild, but we have
various edge cases where we sent the requested start in SnapshotMarker
etc. Within the context of the getItemsForPersistence it seems correct
to update purge even when it is equal to the snapshot start.

Change-Id: If0e7f8da721b474251b888b8ad62ad09288825b5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231404
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>