MB-66315: Improve fuzztest logging
Changed the initialisation logic to avoid excessive logging and to allow
logging to a file. To make this logic reusable, I've pulled it into a
library mcd_fuzzing.

The fuzztest initialization appends an explicit negative filter to the
--gtest_filter argument. The negative filter contains all GTest tests,
and is very long and printed in one large block.

The idea there is that if you specify TestSuite* and that matches
fuzztest and google test tests, and you are running in fuzzing mode, you
only want the fuzzing tests.

However, in our case, when running with CTest, the filter does not have
wildcards and only matches one test exactly. In that scenario, we can
reset the filter to the original value, which also avoids the large
output.

The second issue this addresses is that fuzz tests generate a lot of
output (every time the coverage increases). ctest lacks the facility to
separate the output from multiple test jobs (-j), instead it merges the
output.

When CB_FUZZTEST_LOG_FILE is specified, each fuzztest job will log to a
file with the same name as the executable + fuzztest + the gtest filter
(or timestamp if the filter has wildcards).

Change-Id: I8cf5e1c4f95894786b2f5e47c3c3405e395dc625
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231837
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>