MB-67957: Provide size to inflate snappy
Explicitly provide the max size to snappy for how much we want
to inflate. The value should be guarded on input by checking towards
the max allowed frame size (otherwise it cannot be sent to a client
not supporting compression).

The document will be checked towards the buckets limits later on
and possibly be refused with E2BIG. This change does *not* affect
that logic

Change-Id: Ifa40a35707f4d7474d6bd43148f052399f9c54c5
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/231893
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>