MB-68049: [memcached_testapp] Allow the entire test suite ...
... to be executed in serial mode.

The encryption test suite change the keys and leave files on
disk with keys which wasn't part of the BOOTSTRAP_DEK environment.
This caused the next test suite to fail to start (as it failed
to read configuration and log files).

To work around the problem we should push the current set of
keys every time we start memcached.

The encryption test would also disable encryption on audit
and logs, so the assumption that these files are always encrypted
is now wrong.

Change-Id: I543e8b8019253cc0a453286f6416b7ab722844fc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232156
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dan Owen <owend@couchbase.com>