MB-58132: Add predicate check all items are peristed
In one instance of intermittent test failure, we see only the prepare
was persisted to disk, but not the corresponding the abort yet. The
test goes on to expect the abort when we backfill from disk, but since
that is not persisted, we receive the prepare.

Check the both the prepare & abort are persisted to disk, before
proceeding to backfill.

Change-Id: I3d7b9fa91758c3f2b49a4e3eb22258d15bf8091c
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232309
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>