MB-59438: Disable some Spdlog tests
They fail on neo on the TSan job due to a race. The race is in the
setUpLogger path, which is the test set up, not the race condition that
the test tries to trigger.

The test does not fail on trinity and later, due to series of
refactoring, therefore, it should not be forward merged.

Change-Id: Ia05b68b7ace10ef46df6280593f08871572783dc
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232395
Well-Formed: Restriction Checker
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Vesko Karaganev <vesko.karaganev@couchbase.com>