MB-67668: Fix intermittent XattrTest
The test verifies that after deleting an item with Xattr, the datatype
returned after a get_meta() is XATTR. However, this is only the case if
the get_meta() finds the key before the delete is flushed to disk. If
the get_meta() finds the key after it has been flushed to disk the
datatype will be XATTR and SNAPPY compressed.

Change-Id: If695524284b201b350c5bf182c465f8d76e89a2f
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232501
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>