MB-53543: Re-enable BackfillSmallBuffer and fix
As of https://review.couchbase.org/c/kv_engine/+/180551 / MB-53806, we
no longer set lastReadSeqno before pushing the item over DCP.

The test requires an update to pass, which is to step the backfillPhase
once again, so that lastReadSeqno can be updated as expected.

It is not clear whether this also fixes the intermittent failure we'd
seen before all of those changes.

Change-Id: I61fc0ba451647e1a655d10fb152c88ffcd0e29f0
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232513
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Faizan Alam <faizan.alam@couchbase.com>