MB-68173: fmt::print throws exception for IO errors
As part of adding unit tests for stdin_check we moved from fprintf
to fmt::print due to clang-tidy warnings to move away from the old
C stream interface. Where fprintf would return an error that it
failed to write, fmt would throw an exception we didn't catch.

Note that memcached already ignores SIGPIPE, and we need to do
the same in our test program to avoid the test program to be
terminated due to the signal.

Change-Id: I1886618b2d25faa4990424efc1201f602e91ad32
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232571
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Reviewed-by: Hareen Kancharla <hareen.kancharla@couchbase.com>