MB-61655: Add ConfigurationIface and implement validateParameters
The ConfigurationIface allows validation of a set of parameters. It is
implemented for the ep-engine Configuration, but could be extended to
the memcached Settings in the future (not under this MB).

The validateParameters method is implemented in configuration.cc. It
attempts to set all parameters from the input, and returns suitable
errors on failure. It does not modify the *this object, since it
operates on a copy of it.

The required types for generating a validation map are also defined,
with suitable conversions to JSON, to match what will be sent to the
caller of the validation request (future patch).

Change-Id: I0c772fef94ef5a5cc0bc399c601d16ef97f3e675
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233332
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>