MB-67520: Type correction in StoredValueProperties
The size, metaDataSize and uncompressedSize can produce negative values
(e.g. when reducing the hash-table size via deletion) and hence it's
clearer to keep these as signed types.

Note there is no failure in the current usage as only subtraction and
addition is performed with these values and we do have test coverage.

Change-Id: Ie941b031a81721ec994a162c2f24daa223e5e551
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232910
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Pavlos Georgiou <pavlos.georgiou@couchbase.com>
Well-Formed: Restriction Checker