MB-53189: Add failover-entry based on in-memory state
History:

1. Starting https://review.couchbase.org/c/kv_engine/+/113433, we added
   a failover entry based on the whether if the current snapshot is
   fully persisted, when a replica was promoted to active.
2. Starting https://review.couchbase.org/c/kv_engine/+/113904, we
   persisted the vbucket state transition by adding a meta-item to
   checkpoint, which is persisted to disk when the flusher processes
   that item.

In this commit we are using the lastSnapEndSeqno stored in the
checkpoint manager to evaluate if a snapshot is fully processed in
memory.

When a replica is promoted to active, if a snapshot is fully processed
in memory create the new failover entry, using CM.lastSnapEndSeqno

What happens if the replica has only partially processed a snapshot &
crashes?

When we warmup we check if the last snapshot was fully processed &
accordingly set the lastSnapEndSeqno in CM
(https://review.couchbase.org/c/kv_engine/+/233240/3)

Consider the following example:

1. snapshot [1, 2] fully processed, snapshot [3, 4] partially processed.

   replica: [1, 2] [3, _]
		   ^
		   persistence-cursor
   vbucket-state (on disk): {lastStartSeqno: 2,
			     lastEndSeqno: 4,
		             lastSeqno: 3}
   lastSnapEndSeqno (on CM): 2

- replica crashes & warms-up, we set lastSnapEndSeqno to lastStartSeqno,
  since lastSeqno != lastEndSeqno.

Change-Id: I366d8ea5374990bc2e9caa25f4b5e14b058ddddd
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233181
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>