MB-64001: Add VB:Filter to CacheTransferStream
Three reasons

1) The normal flow will be that a consumer creates a stream and
requests a cache-transfer. When the cache has been transferred
the stream can if needed be "replaced" with an ActiveStream that
will continue regular DCP. That ActiveStream will be given the
filter object that was created in the streamRequest.

2) Later in the CTS patches we will (ab)use the VB::Filter to
allow the consumer to provide more configuration about the CTS,
e.g. by adding more to the stream-request JSON. The VB::Filter
really should now be thought of as a stream config object, it
carries/maintains the stream's collection filter and also various
configuration elements (perhaps to refactor at a later date).

3) Adding a VB::Filter permits the CTS todo privilege checking and
of lesser use be created as a collection filtered stream.

Change-Id: Ibfdd11ac506a16ac71ccd51f46c08fced74bb12b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233097
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Jim Walker <jim@couchbase.com>