MB-64001: Create CacheTransferStream from StreamRequest path
Update the StreamRequest path so that a CacheTransferStream
can be created by the use of a new AddStream flag.

At this patch level, a CacheTransferStream can be created that
will end once the HashTable has been visited (stream ends).
Later changes will allow a CacheTransferStream to complete and
hand-over to an ActiveStream if the consumer request requires
as such.

Change-Id: I3ac612995f4bd82a10c1bb94ace10333759a4a42
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/232444
Tested-by: Jim Walker <jim@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>