MB-53189: Correctly update failover seqno on warmup
We store two snapshot related seqno in the vbucket state on disk -
lastSnapStart, lastSnapEnd.

lastSnapStart is a misnomer though.

1. lastSnapStart is the *end seqno* of the last fully persisted
   snapshot.
2. lastSnapEnd is the *end seqno* of the currently being persisted
   snapshot.

(STParamPersistentBucketTest::CheckFailoverSeqno::CheckFailoverSeqno(),
 confirms the above behavior)

lastSnapStart is update to lastSnapEnd when the current snapshot is
fully persisted to disk.

On warmup, when a vbucket is created, set lastSnapEndSeqno (on CM) to
lastSnapEnd if the snapshot was fully processed, else to lastSnapStart.

This is need in a future patch we rely on CM.lastSnapEndSeqno to create
failover entries when a replica is promoted to active.

Change-Id: I8bba4b295ed727116ed4150e123637e7fabd2a58
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233240
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>