MB-68384: Check for the return value of getenv("HOME")
The common class used to get command line options for the mc tools
have an option to get the JWT information (to ease testing) which was
added in MB-65973.

The problem is that windows do not use HOME as the environment
variable, but HOMEPATH causing getenv to return nullptr and fmt would
crash.

Change-Id: I69c498547beded7cc80870dd867c7abe6863cb68
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233280
Reviewed-by: Vesko Karaganev <vesko.karaganev@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Well-Formed: Restriction Checker