MB-64812: [FBR] Add stats for encryption at rest support
Add stats for ns_server to support encryption at rest with file
based rebalanace.

Add a vBucket filter to the "snapshot-details" group - this will be
used after moving a snapshot to work out which DEKs to import.

Add a "snapshot-deks" stat group - this will be used to report all
DEKs belonging to all snapshots as in-use within ns_server, preventing
their cleanup, and ensuring that the keys are passed down to kv_engine.

Change-Id: Ibb0fa9c0e812b9bef8bed395ce5bbb31df34513b
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233288
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>