MB-61655: Refactor setParameter implementation
SetParameter can be used to change the bucket configuration and to
trigger certain actions in the bucket.

For setting configuration, we can refactor a lot of the code away, and
this patch does just that.

Another weirdness of SetParameter is that each parameter has a category,
and can only be set via that category. To continue to support this, I've
created a list of parameter per category.

When setting configuration via SetParameter, we can now simply lookup
the relevant set of keys for the one we wish to set, and then parse the
value automatically via Configuration::parseAndSetParameter.

This ensures that parameter values are parsed the same between bucket
configuration and SetParameter, and it also eliminates a lot of the
repetitive code which used to parse the value first before setting via
the named accessor.

Finally, this refactor will allow us to add a new "Config" category for
SetParameter, which can be used to set any configuration parameter.

Change-Id: I61dbbd5dcd333f4e535ef3f037c7da3f8f5e3e27
Reviewed-on: https://review.couchbase.org/c/kv_engine/+/233526
Reviewed-by: Trond Norbye <trond.norbye@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>